sdk: moved api client to common package
