+++ sdk/src/main/java/com/sygic/travel/sdk/di/toursModule.kt
@@ -4,7 +4,7 @@ import com.github.salomonbrys.kodein.Kodein
 import com.github.salomonbrys.kodein.bind
 import com.github.salomonbrys.kodein.instance
 import com.github.salomonbrys.kodein.singleton
-import com.sygic.travel.sdk.places.api.SygicTravelApiClient
+import com.sygic.travel.sdk.common.api.SygicTravelApiClient
 import com.sygic.travel.sdk.tours.facade.ToursFacade
 import com.sygic.travel.sdk.tours.service.ToursService
